pyproject.toml
[project]
name = "irs-returns"
version = "0.1.0"
description = "A web application that analyzes public filed returns from tax-exempt organizations using Forms 990, 990-EZ, and 990-PF"
authors = [
    {name = "Jeynald Endaya",email = "jjlendaya@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "dj-database-url (>=3.1.0,<4.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djoser (>=2.3.3,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "redis (>=5.2.0,<6.0.0)",
    "requests (>=2.32.0,<3.0.0)",
    "lxml (>=5.3.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.11,<0.15.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-xdist (>=3.8.0,<4.0.0)",
    "faker (>=40.1.2,<41.0.0)"
]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [
    "*/migrations/*.py",
    "*/migrations/**/*.py"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["core", "organizations", "rest_api", "users", "irs_returns"]
combine-as-imports = true
force-sort-within-sections = true

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "DJ",  # flake8-django
    "N",   # pep8-naming
    "S",   # bandit security
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "S101",  # use of assert (common in tests)
    "E721", # potentially fix and stop ignoring
    "B020", # potentially fix and stop ignoring
    "B023", # potentially fix and stop ignoring
    "B007", # potentially fix and stop ignoring
    "B904", # potentially fix and stop ignoring
    "C401", # potentially fix and stop ignoring
    "C408", # potentially fix and stop ignoring
    "C414", # potentially fix and stop ignoring
    "C405", # potentially fix and stop ignoring
    "C416", # potentially fix and stop ignoring
    "N806", # potentially fix and stop ignoring
]

[tool.ruff.lint.per-file-ignores]
"*/tests/*" = [
  "S101",
  "S105",
  "S106",
  "S108",
  "S311",
]
"*/fixtures/*" = ["S311"]
"*fixture*.py" = ["S311"]
"*/migrations/*" = ["N999"]  # Allow non-snake-case in migrations
